# Optimize _active_tasks