We observed on at least one busy cluster couch_task_status gen_server would not be able to keep up with task updates and `all` task fetching calls. The singleton gen_server would start to get unbounded message backups, where the only solution was to restart couch application.

To help with that there are at least two optimizations we can do:

 1) Avoiding blocking `couch_task_status` gen_sever when fetching all task
properties. That operation now happens outside of the gen_server in the `all` function.

 2) Use a set ets instead of an ordered set ets table. Previously, in CouchDB
1.x, we relied on the ordered ets table to keep a sorted task list, however with clustering in CouchDB 2.x+ we had to call lists:sort/1 on the task result anyway, so there is no reason to waste time maintaining an ordered ets.

To fetch the tasks use the newer `erpc` Erlang module. One concern with this change is how `_active_tasks` will act during rolling upgrades. In that case there are two scenarios:

 1) Old nodes would be calling `gen_server:multi_call/2`. For those nodes we
 left the `all` gen_server callback in so they would not fail.

 2) New nodes calling `erpc:multicall(Nodes, couch_task_status, all, [])` on
 old nodes. In that case the old nodes already have the `all/0` function and
 will in-turn perform a local `gen_server:call(?MODULE, all)` and return the
 results as expected.

While at it, modernize the tests a bit by using the newer `?TDEF_FE` macros and remove optional gen_server callbacks and module attributes.