@nickva 2023-11-16
@rnewson makes sense. Thanks for nudge to keep the concerns in separate in PRs. The vsn PR merged in main. Thanks for the review.