@nickva 2023-11-17
> Did you happen to run any performance benchmarks to demonstrate a performance improvement or a "widening" of the bottleneck?

Good idea. I had setup a custom snippet to create a bunch of active tasks on a node:

```erlang
f(RandProps), RandProps = fun() ->
    [
        {type, indexer},
        {database, <<"foodatabase">>},
        {progress, rand:uniform(100)}
    ]
end.

f(RandTaskProc), RandTaskProc = fun F() ->
    After = 500 + rand:uniform(1000),
    receive
        {add_task, From} ->
            couch_task_status:add_task(RandProps()),
            From ! task_added,
            timer:sleep(1000),
            F()
    after After ->
            couch_task_status:update([{progress, rand:uniform(100)}]),
            F()
    end
end.

f(Spawn), Spawn = fun(N) ->
    Pids = [spawn_link(RandTaskProc) || _ <- lists:seq(1, N)],
    [begin Pid ! {add_task, self()}, receive task_added -> ok end end || Pid <- Pids],
    Pids
end.

f(Pids), Pids = Spawn(400), ok.
```

Then used [vegeta](https://github.com/tsenart/vegeta) to send various rates of requests.

On main:

```
 % echo "GET http://127.0.0.1:15984/_active_tasks" | ./vegeta attack -header "Authorization: Basic YWRtOnBhc3M=" -rate=2000 -duration=30s | ./vegeta report
Requests      [total, rate, throughput]         39178, 1302.25, 791.40
Duration      [total, attack, wait]             42.18s, 30.085s, 12.095s
Latencies     [min, mean, 50, 90, 95, 99, max]  5.734ms, 10.295s, 5.632s, 23.773s, 24.022s, 24.6s, 30.003s
Bytes In      [total, mean]                     2356237741, 60141.86
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           85.20%
Status Codes  [code:count]                      0:5797  200:33381
Error Set:
Get "http://127.0.0.1:15984/_active_tasks": dial tcp 0.0.0.0:0->127.0.0.1:15984: bind: can't assign requested address
Get "http://127.0.0.1:15984/_active_tasks": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
```

Requests started to pile up quickly and some were failing.

With the PR:

```
% echo "GET http://127.0.0.1:15984/_active_tasks" | ./vegeta attack -header "Authorization: Basic YWRtOnBhc3M=" -rate=2000 -duration=30s | ./vegeta report
Requests      [total, rate, throughput]         60000, 2000.03, 1999.80
Duration      [total, attack, wait]             30.003s, 29.999s, 3.551ms
Latencies     [min, mean, 50, 90, 95, 99, max]  2.329ms, 17.76ms, 5.732ms, 43.533ms, 75.153ms, 163.1ms, 442.432ms
Bytes In      [total, mean]                     4246174911, 70769.58
Bytes Out     [total, mean]                     0, 0.00
Success       [ratio]                           100.00%
Status Codes  [code:count]                      200:60000
Error Set:
```

It had no difficulty making 2000 request per second without errors.